Refactor to unexport expander
